feat: Nudge the reverter to leave a comment after reverting
We would like for people to indicate WHY they reverted a PR. Since right
now it's not uncommon to start the work day, see one of your PRs was
reverted, and ask the question "why was this reverted?". You'll have to
track down the person who did the revert, and you'll be lucky if they're
online.

This adds an additional line to the revert comment asking the reverter to
leave a comment describing why the PR was reverted.